+++ R/parse-body.R
@@ -519,6 +519,7 @@ parser_multi <- function() {
     if (!stri_detect_fixed(content_type, "boundary=", case_insensitive = TRUE))
       stop("No boundary found in multipart content-type header: ", content_type)
     boundary <- stri_match_first_regex(content_type, "boundary=([^; ]{2,})", case_insensitive = TRUE)[,2]
+    boundary <- stri_replace_first_regex(boundary, '^"([^"]+)"?', "$1")
     toparse <- parse_multipart(value, boundary)
 
     # set the names of the items as the `name` of each item